Marko

+++ Domain/Orders/OrderedProduct.cs
@@ -12,7 +12,7 @@ namespace Domain.Orders
     {
         public int OrderedProductId { get; set; }
         public int OrderedQuantity { get; set; }
-        public float OrderedPrice { get; set; } //TODO! real(8,2) - 2 digits after comma, 8 digits maximum.
+        public double OrderedPrice { get; set; } //TODO! real(8,2) - 2 digits after comma, 8 digits maximum.
 
         public int OrderId { get; set; }
         public virtual Order Order { get; set; }
