+++ Domain/Store/Product.cs
@@ -14,7 +14,7 @@ namespace Domain.Store
 
         [MaxLength(32)]
         public String ProductName { get; set; } 
-        public float ProductValue { get; set; } //TODO 2 digits after comma, 8 digits max
+        public double ProductValue { get; set; } //TODO 2 digits after comma, 8 digits max
         public bool ProductActive { get; set; }
 
         public int ProductCategoryId { get; set; }
